Added catch and try block to function that gets the eform version from catch and effort JSON
Added class to handle parsing of JSON catch and effrot data from eform version 7.3 and above. The JSON has been optimized to reduce verbosity. Previous JSON was very verbose and could be the cause of out of memory error when these were loaded

Modified API query for downloading JSON from server only for data that has been newly submitted and not yet saved to the database.

Refactoring code adding code that handles optimized catch and effort JSON data in several classes and forms in the application

Added function that identifies if a JSON file is a valid catch and effort JSON history file